Changes the BBC Rarity Rule
This PR adds two new choices to the BBC Rarity rule.

 'Size Matters' which assigns the BBC trait based on the following:

100% chance for mixed, mostly or fully black chars with big_dick_3
50% chance for mixed, mostly or fully black chars with big_dick_2
25% chance for mixed, mostly or fully black chars with big_dick_1

And the 'Black is Better' ruleset which assigns the BBC trait based on the following:
1.0 Multiplier for full black
0.75 multiplier for mostly black
0.5 multiplier for half black
multiplied with the percentages based on the dick-size rule so that a half-black big_dick_1 char has a 25% chance to get the trait, a half-black big_dick_3 has a 50% chance to get the trait, and a full_black big_dick_1 has a 50% chance to get the trait.

The original rule options have been maintained.

Additionally, this adds a decision that only players can use when playing a male, adult character with at least partial mixed race but no BBC trait which grants the BBC trait.

Additionally, this fixes the "Trait Seeding" rule so that it doesn't seed the African trait if the mod is disabled or if trait seeding is set to no.