Fix get_white_vassal_with_non_blacked_wife_scope
get_white_vassal_with_non_blacked_wife_scope differs from the check trigger that goes along with it.

does_white_vassal_has_non_blacked_wife checks for a vassal or liege with a pure wife, but get_white_vassal_with_non_blacked_wife_scope looks for courtiers, liege and lieges courtiers for a man with a pure wife.

This PR changes get_white_vassal_with_non_blacked_wife_scope so that it returns a vassal or liege with a pure wife, making it match does_white_vassal_has_non_blacked_wife and fixing cases of bno_bull.0015 triggering but failing to find valid target1 and target2.

This also fixes an issue with the english localization of bno_bull.0015 where the event referred to the husband as the wife and the wife as the husband. I assume the other translations have the same issue but I'll leave that to the translators.